pyproject.toml
[project]
name = "phi4-multimodal-finetune"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "torch==2.6.0",
    "flash-attn @ https://github.com/Dao-AILab/flash-attention/releases/download/v2.7.4.post1/flash_attn-2.7.4.post1+cu12torch2.6cxx11abiFALSE-cp313-cp313-linux_x86_64.whl",
    "transformers==4.48.2",
    "accelerate==1.3.0",
    "soundfile==0.13.1",
    "pillow==11.1.0",
    "scipy==1.15.2",
    "torchvision==0.21.0",
    "backoff==2.2.1",
    "peft==0.13.2",
    "wandb>=0.16.0",
    "sacrebleu>=2.6.0",
    "datasets>=4.5.0",
    "rouge-score>=0.1.2",
    "scikit-learn>=1.7.2",
    "editdistance>=0.8.1",
]

[tool.uv]
managed = true
